Port BFloat16 scorer fix from NexusMind
Cast logits to float32 before .cpu().numpy() — BFloat16 models
(sustainability_technology v3 + others trained with bf16) would
otherwise raise `TypeError: Got unsupported ScalarType BFloat16`.

Fix was originally in NexusMind 68e3d5d (2026-04-16), but
filter_base_scorer.py lived in both repos and drifted: this repo's
copy lacked the fix. Today's deploy_to_nexusmind.sh propagated the
stale copy, taking production down for ~20 minutes (500s on
/filter/nature_recovery/score) before I caught it.

Separate follow-up: deploy_to_nexusmind.sh should respect a
"NexusMind-owns" list for filter_base_scorer.py and siblings, or
llm-distillery should stop shipping filters/common/ at all. Filing
as a new issue.

+++ filters/common/filter_base_scorer.py
@@ -294,7 +294,10 @@ class FilterBaseScorer(ABC):
 
         with torch.no_grad():
             outputs = self.model(**inputs)
-            raw_scores = outputs.logits[0].cpu().numpy()
+            # .float() before .numpy() — BFloat16 models (sustainability_technology v3,
+            # others trained with bf16) would otherwise raise TypeError: Got unsupported
+            # ScalarType BFloat16. Matches NexusMind 68e3d5d.
+            raw_scores = outputs.logits[0].float().cpu().numpy()
 
         return self._process_raw_scores(raw_scores, result)
 
@@ -385,7 +388,8 @@ class FilterBaseScorer(ABC):
 
                 with torch.no_grad():
                     outputs = self.model(**inputs)
-                    batch_scores = outputs.logits.cpu().numpy()
+                    # See single-article path above — BFloat16 → float32 cast required.
+                    batch_scores = outputs.logits.float().cpu().numpy()
 
                 for j, idx in enumerate(batch_indices):
                     raw_scores = batch_scores[j]
